Add a top-level commit to bio-formats-build bumping submodule version
The current logic of the BIOFORMATS-push works as follows:
- merge all PRs against top-level bio-formats-build
- merge all PRs against submodules
- bump submodules in top-level repository
- bump the component versions across submodules
- push submodule and top-level branches

Although this works fine, this means that the top-level commit was pointing to
the submodule commit prior to the component versions bump and resulted in the
absence of coupling of the components.
This change should include an additional top-level commit implementing this
coupling for downstream jobs.